+++ packages/faux/src/errors.ts
@@ -40,6 +40,8 @@ export function ObjectTypeDoesNotExistError(
   return {
     errorCode: "NOT_FOUND",
     errorName: "ObjectTypeNotFound",
+    errorDescription:
+      "The requested object type is not found, or the client token does not have access to it.",
     errorInstanceId,
     parameters: {
       objectType,
@@ -55,6 +57,8 @@ export function LinkTypeNotFound(
   return {
     errorCode: "NOT_FOUND",
     errorName: "LinkTypeNotFound",
+    errorDescription:
+      "The link type is not found, or the user does not have access to it.",
     errorInstanceId,
     parameters: {
       objectType,
@@ -70,6 +74,8 @@ export function OntologyNotFoundError(
   return {
     errorCode: "NOT_FOUND",
     errorName: "OntologyNotFound",
+    errorDescription:
+      "The requested Ontology is not found, or the client token does not have access to it.",
     errorInstanceId,
     parameters: {
       ontologyRid: ontology,
@@ -85,6 +91,8 @@ export function ObjectNotFoundError(
   return {
     errorCode: "NOT_FOUND",
     errorName: "ObjectNotFound",
+    errorDescription:
+      "The requested object is not found, or the client token does not have access to it.",
     errorInstanceId,
     parameters: {
       objectType,
@@ -97,6 +105,8 @@ export function QueryNotFoundError(queryApiName: string): QueryNotFound {
   return {
     errorCode: "NOT_FOUND",
     errorName: "QueryNotFound",
+    errorDescription:
+      "The query is not found, or the user does not have access to it.",
     errorInstanceId,
     parameters: {
       query: queryApiName,
@@ -108,6 +118,8 @@ export function ActionNotFoundError(): ActionNotFound {
   return {
     errorCode: "NOT_FOUND",
     errorName: "ActionNotFound",
+    errorDescription:
+      "The action is not found, or the user does not have access to it.",
     errorInstanceId,
     parameters: {
       actionRid: "",
@@ -127,6 +139,7 @@ export function InvalidRequest(errorName: string): BaseAPIError {
 export const ApplyActionFailedError: ApplyActionFailed = {
   errorCode: "INVALID_ARGUMENT",
   errorName: "ApplyActionFailed",
+  errorDescription: "",
   errorInstanceId,
   parameters: {},
 };
@@ -134,6 +147,8 @@ export const ApplyActionFailedError: ApplyActionFailed = {
 export const ExecuteQueryFailedError: QueryEncounteredUserFacingError = {
   errorCode: "CONFLICT",
   errorName: "QueryEncounteredUserFacingError",
+  errorDescription:
+    "The authored Query failed to execute because of a user induced error. The message argument is meant to be displayed to the user.",
   errorInstanceId,
   parameters: {
     functionRid:
@@ -146,6 +161,8 @@ export const ExecuteQueryFailedError: QueryEncounteredUserFacingError = {
 export const InvalidContentTypeError: InvalidContentType = {
   errorCode: "INVALID_ARGUMENT",
   errorName: "InvalidContentType",
+  errorDescription:
+    "The Content-Type cannot be inferred from the request content and filename. Please check your request content and filename to ensure they are compatible.",
   errorInstanceId,
   parameters: {},
 };
@@ -153,6 +170,8 @@ export const InvalidContentTypeError: InvalidContentType = {
 export const AttachmentSizeExceededLimitError: AttachmentSizeExceededLimit = {
   errorCode: "INVALID_ARGUMENT",
   errorName: "AttachmentSizeExceededLimit",
+  errorDescription:
+    "The file is too large to be uploaded as an attachment. The maximum attachment size is 200MB.",
   errorInstanceId,
   parameters: {
     fileSizeBytes: "230000",
@@ -163,6 +182,8 @@ export const AttachmentSizeExceededLimitError: AttachmentSizeExceededLimit = {
 export const AttachmentNotFoundError: AttachmentNotFound = {
   errorCode: "NOT_FOUND",
   errorName: "AttachmentNotFound",
+  errorDescription:
+    "The requested attachment is not found, or the client token does not have access to it. Attachments that are not attached to any objects are deleted after two weeks. Attachments that have not been attached to an object can only be viewed by the user who uploaded them. Attachments that have been attached to an object can be viewed by users who can view the object.",
   errorInstanceId,
   parameters: {
     attachmentRid: