+++ packages/client.test.ontology/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/client.test.ontology
 
+## 2.6.0-beta.7
+
+### Patch Changes
+
+- @osdk/api@2.6.0-beta.7
+
+## 2.6.0-beta.6
+
+### Patch Changes
+
+- Updated dependencies [b959085]
+  - @osdk/api@2.6.0-beta.6
+
 ## 2.6.0-beta.5
 
 ### Patch Changes