+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @osdk/widget.api
 
+## 3.3.0-beta.5
+
+### Patch Changes
+
+- @osdk/api@2.6.0-beta.7
+
+## 3.3.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [b959085]
+  - @osdk/api@2.6.0-beta.6
+
 ## 3.3.0-beta.3
 
 ### Patch Changes